all: "Mac OS X" is now called "macOS"
This changes the string we show in the web GUI, a note in the README,
the directory name for the launchd stuff, and prepares the upgrade
system for a future change. The upgrade system learns to recognize
releases named "...-macos-..." in addition to the current
"...-macosx-..." so that we could at some point in the future start
naming our releases like that.