gui: Fix scopes in notification directive (fixes #4941)
The directive lives in its own isolated scope (where we put the
visible() function). The stuff transcluded into the notification
directive lives in the root scope and doesn't have access to the
directive scope. Hence we cannot call dismiss() from *inside* the
directive.

Similarly, `config` does not exist by itself in the directive scope, we
need `$parent` to access the root scope.

Reference: https://docs.angularjs.org/guide/directive#isolating-the-scope-of-a-directive

How this worked before is a mystery. My guess is Angular bug with
directive scope that was fixed in 1.3.