lib/config, lib/model: Limit size of puller queue (fixes #4976)
This adds a new config PullerMaxQueueLength that sets the maximum number
of items we handle in one puller iteration. If set to <= 0 it defaults
to 1000. Point of this is to limit the amount of RAM used for the queue.
Downside is that renames that fall accross batches will not be correctly
detected... Feels more and more like renames should be flagged somehow
on the sending side.