lib/model: Record error for unavailable files
Currently unavailable files show up as out-of-sync in the UI without any progress
or explanation. Now these files will be added to the failed items. The changed
count isn't inreased, as there is no point in retrying those files until a change
in connection happens. At that point, a pull will be scheduled anyway.