lib/db: Use indirect file keys (fixes #5047)
Instead of direcly marshalling the FileInfo for every
(device,folder,file) key we instead go via a layer of indireciton that
is the SHA256 of the marshalled FileInfo. We don't track usage via
refcounts or similar, opting instead for a periodic GC step. This GC
uses a Cuckoo filter to keep track of keys that are (probably) in use,
keeping the memory usage bounded even for large sets of files. A small
number of entries may be left behind due to the probabilistic nature of
the filter, but this is on the order of less than a tenth of a percent
for the largest file sets, less for everyone else.