gui/ignore: sync folder ignores tree view with text rules (fixes #2491)
This commit syncs the treeview and text rules used to ignore
files.

The logic here is to share a common object call
$scope.syncedFiles. The tree view and text rules get synced when we
switch from tab to tab.

When we go from the Advanced (text rules) tab to the Basic (tree view)
tab, we parse the text rules, produce the syncedFiles object and
populate the checkboxes in the tree as needed. A special case is when
the text rules are not parseable (because too complicated), then we
deactivate the tree view. This step is complicated.

When we go from the Basic to the Advanced tab, we produce the shared
syncedFiles object and use that to rewrite the text rules. This step
is simple.