systemd: do not run as root when enabled globally as user service
It is possible to tell systemd to enable a unit for all users in one go which
will configure *user* services for every user:
systemctl --global enable syncthing.service

The problem with that is that root or other system users will then also try to
spawn a syncthing instance on login which is probably not what you want.

Instead only enable the *user* service for normal users.