lib/config: Add encrypted string type, encrypt API key (fixes #5378)
This protects the API key against prying eyes, and more generally
enables storing secrets in the configuration.

When a config entry is set to be an encrypted string it is sealed in a
NaCL secretbox with the SHA256 of our private key as the secret. This
means the secrecy of the secrets in the config is equivalent with the
secrecy of the private key.

Config values can be set in plain text in the config as usual, and are
sent in plain text over the JSON API. This is necessary because the GUI
needs to be able to show the API key. When the config is saved, the
value becomes encrypted. This is indicated by getting a large base64
blob in an "encrypted" attribute. That is,

    <apikey>foo</apikey>

becomes

    <apikey encrypted="..."></apikey>

thus indicating for both humans and machines what's going on.