lib/config, cmd/syncthing: Improve LDAP authentication (fixes #5376)
This adds an extra mode to the LDAP setup, useful for authentication
against AD and enforcing group memberships.

As this requires storing a password in the config, we add support for
encrypted secrets in the config. The password is encrypted using the
hash of our secret key as the key, making the password not readable for
anyone not possessing our secret key.

Additionally the advanced config editor learns to edit the LDAP
configuration.