lib/ur: Implement crash (panic) reporting (fixes #959)
This implements a simple crash reporting method. It piggybacks on the
panic log files created by the monitor process, picking these up and
uploading them from the usage reporting routine.

Internally this is known as "panic uploading" but the user facing term
is "crash reporting" as this is more mainstream.

This becomes a new usage reporting version. I've added strings to the
GUI to the effect that usage reporting includes crash reporting.

The method and server side is trivial; panic logs are uploaded under
their SHA256 hash, with a HEAD request first to avoid sending the entire
log multiple times. Actually doing something useful with this data is up
to the operator of the crash receiver.

A new config value points to the crash receiver base URL, which defaults
to "https://crash.syncthing.net/newcrash" (following the pattern of
"https://data.syncthing.net/newdata" for usage reports, but allowing us
to separate the service as required).