lib/config: Add candidate device list to folder configuration.
Include a list of candidate devices that share the folder, but not
directly with us.  Implemented as a slice internally, and serialized
to XML as repeated candidateDevice tags; to JSON as an array.

A candidate device can be recorded for this folder through the new
AddOrUpdateCandidateDevice() method, optionally with additional info
such as the introducer and its name for the candidate device.  Passed
addresses are collected into a list for possibly trying to contact the
device later.