gui: List suggested devices in folder sharing options.
Include a third category (if non-empty) in the folder sharing device
list.  List the devices collected in the candidateDevices config in
two possible formats:

For already configured devices, show a check mark and the friendly
name; effectively just move entries from the "unrelated" section.

For previously unknown candidate devices, list the complete device ID
as a clickable link, which should open the "Add Device" modal dialog,
but that is not implemented yet [TODO].  If all devices listed under
introducedBy agree on a common name, show that as an alias hint after
the ID.

The lists of suggested devices are temporarily stored within the
currentFolder object, until the folder edit dialog is saved and
closed.

Implement additional {de}selectAllSuggestedDevices() functions that
only operate on the checkmarks, not the unknown devices.