lib/db: Use transaction when flushing batch (fixes #5900)
We use batches when we care about correctness, and try to get the right
write ordering in those batches. But there's nothing to say we don't
shut down (or crash) halfway through writing a batch and end up with
(for example) a FileInfo but no need index, or vice versa (I don't
remember the order but both are bad).

This should make sure than as long as we put all the relevant things in
one batch, that batch should be written atomically.