Update snapcraft.yaml.template
Adding a `base` stanza will cause snapcraft to use newer (supported) functionality. Typically this would mean running `snapcraft` will assume you have `multipass` installed (see https://multipass.run/) inside which the snap build will run. However, in a CI environment, multipass (which uses kvm) will sometimes fail, as many CI systems won't allow a "nested" VM environment. So there's a flag for `snapcraft` called `--destructive-mode` which means 'I know what I'm doing, build on the host I'm running snapcraft on, not in any additional container or VM, thanks'. Which is effectively similar to the legacy way of running snapcraft on the host. The reason we prefer not to do that is that people make the mistake of running snapcraft on the 'wrong' release of Ubuntu, which can result in linker errors or binaries built for an older (or newer) libc. The resulting snap may well work locally, but will likely fail once delivered to a user system, running some other random release.