gui: Prevent text overflow in file lists (fixes #6268)
Allow to break words in long filenames and paths listed in "Out of Sync
Items", "Recent Changes", and "Failed Items". This prevents text overflow
when there are no spaces in file names, which are too long to fit on the
screen.

word-break: break-all and not word-break: break-word is needed to force
breakage of long names with no legible words in them, and also to break
CJK text. In addition, word-break: break-word is not supported by IE.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>