lib/db: Add version vector indirection (ref #6372)
This adds indirection of large version vectors in the same manner as we
already to block lists. The effect is the same: less duplicated data in
some situations.

To mitigate the impact for when this indirection
wouldn't be needed I've added an indirection cutoff for both blocks and
the new version vector stuff: we don't do the indirection at all for
small block lists or small version vectors, instead storing it directly
like we used to do. This is faster for small files and small setups.

I also changed so that the putFile always recalculates the hash key.
This is paranoia in the case of the block lists, but fairly harmless
paranoia that avoids the chaos that could ensue if we were fed items
with the wrong hash key. For the version vector, we have code that just
updates the version and not much else on the file, so we need to make
sure on the way in that the hash is correct anyway.