lib/db: Unify error handling in PendingDevices().
Handle several possible errors in the call sequence by jumping to a
common label within the iterator loop, instead of collecting error
artifacts and checking at the end.

This is arguably more readable by using the goto construct.
Apparently it has a better reputation in Go than most other
programming languages, for a limited set of valid use-cases.  This is
such a case in my opinion, with very narrow scope and leading to more
readable code.