lib/model: Clean out pending stuff on config changes.
When the configuration is changed, so far it takes care by itself to
not keep pendingFolder or pendingDevice entries which are either
shared or ignored in the new state.  The corresponding database
entries need to be cleaned out by the model though, when informed
about config changes in CommitConfiguration().

Also clean out the entries when starting up with a new configuration
in onServe().